moved crud functionality for medication card to higher parent component; no longer repeating on each list page
