+++ src/ApplicationViews.js
@@ -1,5 +1,9 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import { Route, Redirect } from "react-router-dom";
+import ApplicationManager from "./components/modules/ApplicationManager";
+import { currentDateTime } from "./components/modules/helperFunctions";
+import { calculateNextRefill } from "./components/modules/helperFunctions";
+import { useHistory } from "react-router-dom";
 
 //login
 import Login from "./components/auth/Login";
@@ -25,8 +29,8 @@ import EditPharmacyForm from "./components/pharmacy/EditPharmacyForm";
 
 
 const ApplicationViews = (props) => {
-    const hasUser = props.hasUser
-    const setUser = props.setUser
+    const { hasUser, setUser, removeDrug, drugs, setDrugs, handleChange, getIdOfDrug, drug, drugImage, handlePharmacyDropdown, pharmacyList, uploadImage, handleFieldChange, handleDrugForm, toggle, modal, toggleNested, toggleAll, nestedModal, closeAll } = props
+
     return (
         <>
             <Route path="/login"
@@ -42,19 +46,18 @@ const ApplicationViews = (props) => {
 
             <Route exact path="/"
                 render={props => {
-                    return (hasUser ? <Home hasUser={hasUser} {...props} drugId={parseInt(props.match.params.drugId)} /> : <Redirect to="/login" />)
+                    return (hasUser ? <Home hasUser={hasUser} {...props} drugId={parseInt(props.match.params.drugId)} setDrugs={setDrugs} drugs={drugs} removeDrug={removeDrug}
+                        drug={drug} handlePharmacyDropdown={handlePharmacyDropdown} pharmacyList={pharmacyList} drugImage={drugImage} uploadImage={uploadImage} handleFieldChange={handleFieldChange} handleDrugForm={handleDrugForm}
+                        nestedModal={nestedModal} toggle={toggle} modal={modal} toggleNested={toggleNested} toggleAll={toggleAll} closeAll={closeAll} getIdOfDrug={getIdOfDrug} handleChange={handleChange}
+                    /> : <Redirect to="/login" />)
                 }}
             />
 
-            {/* <Route exact path="/medication/new"
-                render={props => {
-                    return <AddMedicationFormModal {...props} />
-                }}
-            /> */}
-
             <Route exact path="/medication/list"
                 render={props => {
-                    return (hasUser ? <MedicationList drugId={parseInt(props.match.params.drugId)} {...props} /> : <Redirect to="/login" />)
+                    return (hasUser ? <MedicationList drugId={parseInt(props.match.params.drugId)} {...props} removeDrug={removeDrug} drugs={drugs} drug={drug} handlePharmacyDropdown={handlePharmacyDropdown} pharmacyList={pharmacyList} drugImage={drugImage} uploadImage={uploadImage} handleFieldChange={handleFieldChange} handleDrugForm={handleDrugForm}
+                        nestedModal={nestedModal} toggle={toggle} modal={modal} toggleNested={toggleNested} toggleAll={toggleAll} closeAll={closeAll} getIdOfDrug={getIdOfDrug} handleChange={handleChange} setDrugs={setDrugs}
+                    /> : <Redirect to="/login" />)
                 }}
             />
