+++ src/components/medication/styles/MedicationCard.css
@@ -44,12 +44,6 @@
   padding: 0.5em;
 }
 
-/* .medicationCard {
-  background-color: #acbea380;
-  box-shadow: 1px 1px 1px inset;
-  margin: 1.5em;
-} */
-
 .medicationCard--background-green {
   height: 100%;
   background-color: #acbea380;
